python3/16_Web_Services/c_web_application/using_flask/new/requirements.txt
flask==2.0.2
pyjwt[crypto]==2.3.0
requests==2.26.0
requests-oauthlib==1.3.0
selenium==4.1.0
